修复 papers100M GPU OOM: int32 CSR 加载 + 修正分区数
1. appnp_backend._load_csr_graph: rowptr/col 直接以 int32 加载到 GPU
   - papers100M col 1.6B edges: 12.8 GB(int64) → 6.4 GB(int32), 省 ~6.4 GB
   - 111M nodes < 2^31, int32 完全安全

2. 修正 n_parts 与其他数据集对齐 (~10-30K nodes/window):
   - GPH_Slim:  64→4096 (~31K/window, ~1.1 GB GPU)
   - GT:        64→4096 (~31K/window, ~2.0 GB GPU)
   - GPH_Large: 512→8192 (~16K/window, ~9.5 GB GPU)

+++ NeutronGT/core/ppr_backends/appnp_backend.py
@@ -22,8 +22,10 @@ def _load_csr_graph(csr_data, device, num_nodes: int | None = None):
     if "rowptr" not in csr_data or "col" not in csr_data:
         raise KeyError(f"CSR data must contain rowptr and col, got keys={list(csr_data.keys())}")
 
-    rowptr = torch.as_tensor(csr_data["rowptr"], dtype=torch.long, device=device)
-    col = torch.as_tensor(csr_data["col"], dtype=torch.long, device=device)
+    # 大图使用 int32 加载并直接放在 GPU 上，相比 int64 节省一半显存
+    # papers100M: 1.6B edges * 4 bytes = 6.4 GB (int32) vs 12.8 GB (int64)
+    rowptr = torch.as_tensor(csr_data["rowptr"], dtype=torch.int32, device=device)
+    col = torch.as_tensor(csr_data["col"], dtype=torch.int32, device=device)
     if rowptr.dim() != 1 or col.dim() != 1:
         raise ValueError("CSR rowptr and col must be 1-D")
     inferred_num_nodes = int(rowptr.numel() - 1)
